+++ awswrangler/s3/_read_text.py
@@ -3,7 +3,7 @@ import datetime
 import itertools
 import logging
 import pprint
-from typing import Any, Callable, Dict, Iterator, List, Optional, Union
+from typing import Any, Callable, Dict, Iterator, List, Literal, Optional, Union, overload
 
 import boto3
 import pandas as pd
@@ -146,8 +146,52 @@ def _read_text_format(
     )
 
 
+@overload
 def read_csv(
     path: Union[str, List[str]],
+    chunksize: None = ...,
+    path_suffix: Union[str, List[str], None] = ...,
+    path_ignore_suffix: Union[str, List[str], None] = ...,
+    version_id: Optional[Union[str, Dict[str, str]]] = ...,
+    ignore_empty: bool = ...,
+    use_threads: Union[bool, int] = ...,
+    last_modified_begin: Optional[datetime.datetime] = ...,
+    last_modified_end: Optional[datetime.datetime] = ...,
+    boto3_session: Optional[boto3.Session] = ...,
+    s3_additional_kwargs: Optional[Dict[str, Any]] = ...,
+    dataset: bool = ...,
+    partition_filter: Optional[Callable[[Dict[str, str]], bool]] = ...,
+    parallelism: int = ...,
+    **pandas_kwargs: Any,
+) -> pd.DataFrame:
+    ...
+
+
+@overload
+def read_csv(
+    path: Union[str, List[str]],
+    chunksize: int,
+    path_suffix: Union[str, List[str], None] = ...,
+    path_ignore_suffix: Union[str, List[str], None] = ...,
+    version_id: Optional[Union[str, Dict[str, str]]] = ...,
+    ignore_empty: bool = ...,
+    use_threads: Union[bool, int] = ...,
+    last_modified_begin: Optional[datetime.datetime] = ...,
+    last_modified_end: Optional[datetime.datetime] = ...,
+    boto3_session: Optional[boto3.Session] = ...,
+    s3_additional_kwargs: Optional[Dict[str, Any]] = ...,
+    dataset: bool = ...,
+    partition_filter: Optional[Callable[[Dict[str, str]], bool]] = ...,
+    parallelism: int = ...,
+    **pandas_kwargs: Any,
+) -> Iterator[pd.DataFrame]:
+    ...
+
+
+@overload
+def read_csv(
+    path: Union[str, List[str]],
+    chunksize: Optional[int] = None,
     path_suffix: Union[str, List[str], None] = None,
     path_ignore_suffix: Union[str, List[str], None] = None,
     version_id: Optional[Union[str, Dict[str, str]]] = None,
@@ -157,7 +201,6 @@ def read_csv(
     last_modified_end: Optional[datetime.datetime] = None,
     boto3_session: Optional[boto3.Session] = None,
     s3_additional_kwargs: Optional[Dict[str, Any]] = None,
-    chunksize: Optional[int] = None,
     dataset: bool = False,
     partition_filter: Optional[Callable[[Dict[str, str]], bool]] = None,
     parallelism: int = 200,
